XECS-480, XECS-890 Merged revisions 12561-12568,12570-12572,12574 via svnmerge from https://sipxecs.sipfoundry.org/rep/sipXecs/branches/64bit-rebased-2
........
  r12561 | dkrzemin | 2008-05-06 12:58:58 -0400 (Tue, 06 May 2008) | 3 lines
  
  All the changes in 64bit branch until fatal merge from 3.10 branch.
........
  r12562 | dkrzemin | 2008-05-06 12:59:17 -0400 (Tue, 06 May 2008) | 3 lines
  
  Re-enable -Werror on 64bit branch, and fix one warning which crept in
........
  r12563 | dkrzemin | 2008-05-06 12:59:23 -0400 (Tue, 06 May 2008) | 3 lines
  
  Fix 64bit warnings in osbrec.cpp
........
  r12564 | dkrzemin | 2008-05-06 12:59:28 -0400 (Tue, 06 May 2008) | 6 lines
  
  XECS-480 - Port sipXecs to 64bit SUSE
  
  Small whitespace changes.
  <ke>
........
  r12565 | dkrzemin | 2008-05-06 12:59:39 -0400 (Tue, 06 May 2008) | 10 lines
  
  XECS-480 - Port sipXecs to 64bit SUSE
  
  - Additional intptr_t for consistency.
  - Replace '#' in comments with [pound] to clear compiler warnings (2 files).
  - In sipXportLib/src/test/os/OsProcessTest.cpp:testLaunch(), disable priority
    ASSERT using KNOWN_BUG because it was intermittently failing.
  - More whitespace changes.
  <ke>
........
  r12568 | keccles | 2008-05-07 10:13:29 -0400 (Wed, 07 May 2008) | 6 lines
  
  XECS-480 - Port sipXecs to 64bit SUSE 
  
  Check-in some fixes to 64bit-rebased-2 branch so latest main changes can be
  merged to the branch.  This version DOES NOT BUILD.
  <ke>
........
  r12572 | keccles | 2008-05-07 16:09:01 -0400 (Wed, 07 May 2008) | 4 lines
  
  XECS-480 XECS-890
  One more 64-bit fix before merge to main.
........


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12576 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXacd/src/ACDServer.h
@@ -126,7 +126,7 @@ public:
 
    int                getAdministrativeState(void);
    
-   int                getMaxCallAllowed(void) { return mMaxAcdCallsAllowed; }
+   size_t             getMaxCallAllowed(void) { return mMaxAcdCallsAllowed; }
 
 /* ============================ INQUIRY =================================== */
