+++ sipXcallLib/examples/sipXtapiDriver/include/sipXtapiDriver/sipXtapi.h
@@ -276,7 +276,7 @@ typedef unsigned int SIPX_CONF ;
  * automatically deallocate this handle immediately after the status
  * call back.
  */
-typedef unsigned int SIPX_INFO;
+typedef intptr_t SIPX_INFO;
 
 
 /**
@@ -292,7 +292,7 @@ typedef unsigned int SIPX_PUB;
  * SIPX_SUB handles are created by using the sipxCallSubscribe function.
  * SIPX_SUB handles should be destroyed using the sipxCallUnsubscribe function.
  */
-typedef unsigned int SIPX_SUB ;
+typedef intptr_t SIPX_SUB ;
 
 /** 
  * A handle referring to an SIP NOTIFY message.
